+++ src/main/java/com/example/project4/controller/AdminController.java
@@ -50,6 +50,7 @@ public class AdminController {
 @GetMapping(value="/notificationBoard/{notificationId}")
     public String notificationDetail(@PathVariable("notificationId") Long notificationId, Model model) {
         try{
+            System.out.println("실행됨");
             NotificationFormDto notificationFormDto = boardService.notificationDetail(notificationId);
             /* 게시글 조회 시*/
             /* 방문자의 경우 좋아요 표시 빈하트 */
@@ -102,18 +103,5 @@ public class AdminController {
         boardService.deleteBoard(notificationId);
 
         return "redirect:/";
-    }
-    /* 좋아요 클릭시 */
-    @PostMapping(value="/notificationBoard/{notificationId}/like")
-    @ResponseBody
-    public String like(@RequestBody Map<String, Long> requestBody, @PathVariable("notificationId") Long notificationId) {
-        /* 미완성, 현재 좋아요 클릭시 게시글 좋아요 카운트 기능만 완성함. */
-        /* 멤버 기능 완성되면, 로그인한 유저가 해당 게시글 좋아요 했는지 여부 Boolean으로 반환하기 */
-        /* 반환 후 boolean 값에 따라 제이 쿼리 사용하여 좋아요 아이콘 하트 또는 빈 하트로 나오게 수정. */
-
-        Long data = requestBody.get("data");
-
-        Long like = boardService.like_count(data, notificationId);
-        return like.toString();
     }
 }
