+++ src/main/java/com/example/project4/entity/BaseEntity.java
@@ -21,4 +21,5 @@ public abstract class BaseEntity extends BaseTimeEntity{
     @LastModifiedBy
     private String modifiedBy;
 
+
 }
