+++ src/main/resources/application.properties
@@ -9,5 +9,11 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
 
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+
+spring.thymeleaf.mode=HTML
+spring.thymeleaf.cache=false
 spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html   
+spring.thymeleaf.suffix=.html
+spring.thymeleaf.encoding=UTF-8
+spring.thymeleaf.content-type=text/html
+spring.thymeleaf.extras.java8time.enabled=true
