+++ src/main/resources/templates/board/notice_new_board.html
@@ -15,14 +15,15 @@
         <div class="form_data">제목 <input type="text" name="title" id="title" th:field="*{title}"></div>
         <div class="form_data content">내용 <input type="text" name="content" id="content" th:field="*{content}"></div>
         <input type="submit" value="등록하기"/>
+        <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
     </form>
 
-    <h2 th:if="${board}==update_notice" class="head_title">공지사항 수정페이지</h2>
+<!--    <h2 th:if="${board}==update_notice" class="head_title">공지사항 수정페이지</h2>
     <form th:if="${board}==update_notice" th:action="@{'/admin/notificationBoard/' + ${notificationFormDto.notificationId}}" method="post" th:object="${notificationFormDto}">
         <div class="form_data">제목 <input type="text" name="title" id="title" th:field="*{title}" th:value="${notificationFormDto.getTitle()}"></div>
         <div class="form_data content">내용 <input type="text" name="content" id="content" th:field="*{content}" th:value="${notificationFormDto.getContent()}"></div>
         <input type="submit" value="수정하기"/>
-    </form>
+    </form>-->
 </div>
 
 </div>
